User Model, Authentication Logic
I have implemented authentication logic to the Events controller. DIDN'T HAVE THE CHANCE TO TEST IT YET, but I will as soon as I come back.

I have created a user model that has an ID field and one to many relationship to Events, Venues and People.

+++ KonneyTM/Controllers/EventsController.cs
@@ -1,9 +1,12 @@
 ﻿using KonneyTM.DAL;
+using KonneyTM.Models;
 using KonneyTM.ViewModels;
+using Microsoft.AspNet.Identity;
 using System;
 using System.Collections.Generic;
 using System.IO;
 using System.Linq;
+using System.Security.Authentication;
 using System.Web;
 using System.Web.Mvc;
 
@@ -13,7 +16,17 @@ namespace KonneyTM.Controllers
     {
         public ActionResult Index()
         {
-            var eventVMs = EventViewModel.GetAllAsOrderedList();
+            List<EventViewModel> eventVMs;
+
+            if (User.Identity.IsAuthenticated)
+            {
+                eventVMs = EventViewModel.GetAll(User.Identity.GetUserId());
+            }
+            else
+            {
+                eventVMs = EventViewModel.GetAll("demo");
+            }
+
             return View(eventVMs);
         }
         
@@ -32,10 +45,20 @@ namespace KonneyTM.Controllers
                 string extension = Path.GetExtension(eventVM.ImageFile.FileName);
                 string imageFileName = DateTime.Now.ToString("yyyyMMddHHmmss") + extension;
                 eventVM.ImagePath = imageFileName;
-                imageFileName = Path.Combine(Server.MapPath("~/Images/Events"), imageFileName);
-                eventVM.ImageFile.SaveAs(imageFileName);
+                
+                if (User.Identity.IsAuthenticated)
+                {
+                    imageFileName = Path.Combine(Server.MapPath($"~/Images/{User.Identity.GetUserId()}/Events"), imageFileName);
+                    eventVM.ImageFile.SaveAs(imageFileName);
+                    eventVM.SaveToDB(User.Identity.GetUserId());
+                }
+                else
+                {
+                    imageFileName = Path.Combine(Server.MapPath($"~/Images/demo/Events"), imageFileName);
+                    eventVM.ImageFile.SaveAs(imageFileName);
+                    eventVM.SaveToDB("demo");
+                }
 
-                eventVM.SaveToDB();
                 return RedirectToAction("Index");
             }
 
@@ -45,9 +68,32 @@ namespace KonneyTM.Controllers
         public ActionResult Event(int id)
         {
             using (var db = new KonneyContext())
-            { 
-                var eventVM = EventViewModel.FromEvent(db.Events.First(e => e.ID == id));
-                return View(eventVM);
+            {
+                var ev = db.Events.First(e => e.ID == id);
+                
+                if (User.Identity.IsAuthenticated)
+                {
+                    string userID = User.Identity.GetUserId();
+
+                    if (userID == ev.User.ID)
+                    {
+                        var eventVM = EventViewModel.FromEvent(ev);
+                        return View(eventVM);
+                    }
+                    else
+                    {
+                        throw new AuthenticationException("You are not authorized to edit this event.");
+                    }
+                }
+                else if (ev.User.ID == "demo")
+                {
+                    var eventVM = EventViewModel.FromEvent(ev);
+                    return View(eventVM);
+                }
+                else
+                {
+                    throw new Exception("Something went wrong...");
+                }
             }
         }
 
@@ -56,18 +102,55 @@ namespace KonneyTM.Controllers
         {
             if (ModelState.IsValid)
             {
-                if (eventVM.ImageFile != null)
+                using (var context = new KonneyContext())
                 {
-                    //Image Upload Logic
-                    string extension = Path.GetExtension(eventVM.ImageFile.FileName);
-                    string imageFileName = DateTime.Now.ToString("yyyyMMddHHmmss") + extension;
-                    eventVM.ImagePath = imageFileName;
-                    imageFileName = Path.Combine(Server.MapPath("~/Images/Events"), imageFileName);
-                    eventVM.ImageFile.SaveAs(imageFileName);
-                }
+                    //Get the related event
+                    var ev = context.Events.Single(e => e.ID == eventVM.ID);
 
-                eventVM.SubmitChanges();
-                return RedirectToAction("Event", new { id = eventVM.ID });
+                    //If the user is logged in
+                    if (User.Identity.IsAuthenticated)
+                    {
+                        //If the event belongs to the user
+                        if (ev.User.ID == User.Identity.GetUserId()) 
+                        {
+                            //If an image file is selected
+                            if (eventVM.ImageFile != null)
+                            {
+                                string extension = Path.GetExtension(eventVM.ImageFile.FileName);
+                                string imageFileName = DateTime.Now.ToString("yyyyMMddHHmmss") + extension;
+                                eventVM.ImagePath = imageFileName;
+                                imageFileName = Path.Combine(Server.MapPath($"~/Images/{User.Identity.GetUserId()}/Events"), imageFileName);
+                                eventVM.ImageFile.SaveAs(imageFileName);
+                            }
+
+                            eventVM.SubmitChanges();
+                            return RedirectToAction("Event", new { id = eventVM.ID });
+                        }
+                        else
+                        {
+                            throw new AuthenticationException("You are not authorized to edit this event.");
+                        }
+                    }
+                    //If the application is in demo mode
+                    else if (ev.User.ID == "demo")
+                    {
+                        if(eventVM.ImageFile != null)
+                        {
+                            string extension = Path.GetExtension(eventVM.ImageFile.FileName);
+                            string imageFileName = DateTime.Now.ToString("yyyyMMddHHmmss") + extension;
+                            eventVM.ImagePath = imageFileName;
+                            imageFileName = Path.Combine(Server.MapPath($"~/Images/demo/Events"), imageFileName);
+                            eventVM.ImageFile.SaveAs(imageFileName);
+                        }
+
+                        eventVM.SubmitChanges();
+                        return RedirectToAction("Event", new { id = eventVM.ID });
+                    }
+                    else
+                    {
+                        throw new Exception("Something went wrong...");
+                    }
+                }
             }
             return RedirectToAction("Event", new { id = eventVM.ID });
         }
@@ -77,25 +160,92 @@ namespace KonneyTM.Controllers
             using (var db = new KonneyContext())
             {
                 var ev = db.Events.First(e => e.ID == id);
-                db.Events.Remove(ev);
-                db.SaveChanges();
+
+                if (User.Identity.IsAuthenticated)
+                {
+                    string userID = User.Identity.GetUserId();
+
+                    if (userID == ev.User.ID)
+                    {
+                        db.Events.Remove(ev);
+                        db.SaveChanges();
+                        return RedirectToAction("Index");
+                    }
+                    else
+                    {
+                        throw new AuthenticationException("You are not authorized to delete this event.");
+                    }
+                }
+                else if (ev.User.ID == "demo")
+                {
+                    db.Events.Remove(ev);
+                    db.SaveChanges();
+                }
             }
-            return RedirectToAction("Index");
+
+            throw new Exception("Something went wrong...");
         }
 
         public ActionResult ChangeVenue(int eventID)
         {
-            var changeVenue = new ChangeVenueVM { EventID = eventID };
-            return View(changeVenue);
+            using (var context = new KonneyContext())
+            {
+                var relatedEvent = context.Events.Single(e => e.ID == eventID);
+                
+                if(User.Identity.IsAuthenticated)
+                {
+                    var userID = User.Identity.GetUserId();
+
+                    if(relatedEvent.User.ID == userID)
+                    {
+                        var changeVenue = new ChangeVenueVM { EventID = eventID };
+                        return View(changeVenue);
+                    }
+                    else
+                    {
+                        throw new AuthenticationException("You are not authorized to change the venue of this event.");
+                    }
+                }
+                else if(relatedEvent.User.ID == "demo")
+                {
+                    var changeVenue = new ChangeVenueVM { EventID = eventID };
+                    return View(changeVenue);
+                }
+                else
+                {
+                    throw new Exception("Something went wrong.");
+                }
+            }
         }
 
         public ActionResult SubmitVenueChange(int eventID, int venueID)
         {
             using (var db = new KonneyContext())
             {
-                var ev = db.Events.First(e => e.ID == eventID);
-                ev.Place = db.Venues.First(v => v.ID == venueID);
-                db.SaveChanges();
+                var relatedEvent = db.Events.First(e => e.ID == eventID);
+
+                if (User.Identity.IsAuthenticated)
+                {
+                    var userID = User.Identity.GetUserId();
+                    if (relatedEvent.User.ID == userID)
+                    {
+                        relatedEvent.Place = db.Venues.First(v => v.ID == venueID);
+                        db.SaveChanges();
+                    }
+                    else
+                    {
+                        throw new AuthenticationException("You are not authorized to change the venue of this event.");
+                    }
+                }
+                else if (relatedEvent.User.ID == "demo")
+                {
+                    relatedEvent.Place = db.Venues.First(v => v.ID == venueID);
+                    db.SaveChanges();
+                }
+                else
+                {
+                    throw new Exception("Something went wrong...");
+                }
             }
 
             return RedirectToAction("Event", new { id = eventID });
@@ -105,19 +255,53 @@ namespace KonneyTM.Controllers
         {
             using (var db = new KonneyContext())
             {
-                var addPerson = new AddPersonVM { EventID = eventID };
-                var relatedEvent = EventViewModel.FromEvent(db.Events.First(e => e.ID == eventID));
-                var allPeople = PersonViewModel.GetAllAsOrderedList();
+                var relatedEvent = db.Events.First(e => e.ID == eventID);
 
-                foreach (var p in allPeople)
+                if (User.Identity.IsAuthenticated)
                 {
-                    if (!relatedEvent.InvitedPeopleIDs.Contains(p.ID))
+                    var userID = User.Identity.GetUserId();
+
+                    if (relatedEvent.User.ID == userID)
                     {
-                        addPerson.People.Add(p);
+                        var addPerson = new AddPersonVM { EventID = eventID };
+                        var allPeople = PersonViewModel.GetAllAsOrderedList();
+                        var eventVM = EventViewModel.FromEvent(relatedEvent);
+
+                        foreach (var p in allPeople)
+                        {
+                            if (!eventVM.InvitedPeopleIDs.Contains(p.ID))
+                            {
+                                addPerson.People.Add(p);
+                            }
+                        }
+
+                        return View(addPerson);
+                    }
+                    else
+                    {
+                        throw new AuthenticationException("You are not authorized to add people to this event.");
                     }
                 }
-                
-                return View(addPerson);
+                else if (relatedEvent.User.ID == "demo")
+                {
+                    var addPerson = new AddPersonVM { EventID = eventID };
+                    var allPeople = PersonViewModel.GetAllAsOrderedList();
+                    var eventVM = EventViewModel.FromEvent(relatedEvent);
+
+                    foreach (var p in allPeople)
+                    {
+                        if (!eventVM.InvitedPeopleIDs.Contains(p.ID))
+                        {
+                            addPerson.People.Add(p);
+                        }
+                    }
+
+                    return View(addPerson);
+                }
+                else
+                {
+                    throw new Exception("Something went wrong.");
+                }
             }
         }
 
@@ -125,11 +309,35 @@ namespace KonneyTM.Controllers
         {
             using (var db = new KonneyContext())
             { 
-                var relatedEvent = EventViewModel.FromEvent(db.Events.First(e => e.ID == eventID));
-                var relatedPerson = PersonViewModel.FromPerson(db.People.First(p => p.ID == personID));
-                relatedEvent.InvitedPeopleIDs.Add(relatedPerson.ID);
-                relatedEvent.SubmitChanges();
+                var relatedEvent = db.Events.First(e => e.ID == eventID);
+                var relatedPerson = db.People.First(p => p.ID == personID);
+                var eventVM = EventViewModel.FromEvent(relatedEvent);
+
+                if (User.Identity.IsAuthenticated)
+                {
+                    var userID = User.Identity.GetUserId();
+
+                    if(userID == relatedEvent.User.ID && userID == relatedPerson.User.ID)
+                    {
+                        eventVM.InvitedPeopleIDs.Add(relatedPerson.ID);
+                        eventVM.SubmitChanges();
+                    }
+                    else
+                    {
+                        throw new AuthenticationException("You are not authorized to add people to this event.");
+                    }
+                }
+                else if (relatedPerson.User.ID == "demo" && relatedEvent.User.ID == "demo")
+                {
+                    eventVM.InvitedPeopleIDs.Add(relatedPerson.ID);
+                    eventVM.SubmitChanges();
+                }
+                else
+                {
+                    throw new Exception("Something went wrong.");
+                }
             }
+
             return RedirectToAction("Event", new { id = eventID });
         }
 
@@ -137,10 +345,33 @@ namespace KonneyTM.Controllers
         {
             using (var db = new KonneyContext())
             { 
-                var relatedEvent = EventViewModel.FromEvent(db.Events.First(e => e.ID == eventID));
-                relatedEvent.InvitedPeopleIDs.RemoveAll(i => i == personID);
-                relatedEvent.SubmitChanges();
+                var relatedEvent = db.Events.First(e => e.ID == eventID);
+                var eventVM = EventViewModel.FromEvent(relatedEvent);
+
+                if (User.Identity.IsAuthenticated)
+                {
+                    var userID = User.Identity.GetUserId();
+                    if (relatedEvent.User.ID == userID)
+                    {
+                        eventVM.InvitedPeopleIDs.RemoveAll(i => i == personID);
+                        eventVM.SubmitChanges();
+                    }
+                    else
+                    {
+                        throw new AuthenticationException("You are not authorized to remove people from this event.");
+                    }
+                }
+                else if (relatedEvent.User.ID == "demo")
+                {
+                    eventVM.InvitedPeopleIDs.RemoveAll(i => i == personID);
+                    eventVM.SubmitChanges();
+                }
+                else
+                {
+                    throw new Exception("Something went wrong");
+                }
             }
+
             return RedirectToAction("Event", new { id = eventID });
         }
     }
